requirements_cpu.txt
#
# python  (3.8.5)
# pip     (21.1.1)
#
# author: @contimatteo
#

#############################
## Versioned

# https://discourse.aicrowd.com/t/unable-to-install-flatland/1391/7
# importlib-metadata==0.23

flatland-rl==2.2.2
tensorflow
tensorforce==0.6.5

#############################
## Not-Versioned

keras-rl
scikit-learn

#############################
## Dev

pylint
yapf

#############################
## OSX Big Sur compatibility
## https://github.com/openai/gym/issues/2101

# pyglet==1.5.11
